+++ src/DSerfozo.RpcBindings.CefGlue/Common/Serialization/RpcMessageSerializer.cs
@@ -19,16 +19,16 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
             {
                 message = CefProcessMessage.Create(Messages.DynamicObjectRequestMessageName);
                 var arguments = message.Arguments;
-                arguments.SetInt(0, dynamicObjectRequest.ExecutionId);
+                arguments.SetInt64(0, dynamicObjectRequest.ExecutionId);
                 arguments.SetString(1, dynamicObjectRequest.Name);
             }
             else if (methodExecution != null)
             {
                 message = CefProcessMessage.Create(Messages.MethodExecutionMessageName);
                 var arguments = message.Arguments;
-                arguments.SetInt(0, methodExecution.ExecutionId);
-                arguments.SetInt(1, methodExecution.MethodId);
-                arguments.SetInt(2, methodExecution.ObjectId);
+                arguments.SetInt64(0, methodExecution.ExecutionId);
+                arguments.SetInt64(1, methodExecution.MethodId);
+                arguments.SetInt64(2, methodExecution.ObjectId);
                 using (var list = CefListValue.Create())
                 {
                     for (var i = 0; i < methodExecution.Parameters.Length; i++)
@@ -44,7 +44,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                 message = CefProcessMessage.Create(Messages.CallbackResultMessageName);
                 var arguments = message.Arguments;
 
-                arguments.SetInt(0, callbackResult.ExecutionId);
+                arguments.SetInt64(0, callbackResult.ExecutionId);
                 arguments.SetBool(1, callbackResult.Success);
                 if (!callbackResult.Success)
                 {
@@ -70,7 +70,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
             {
                 message = CefProcessMessage.Create(Messages.MethodResultMessageName);
                 var arguments = message.Arguments;
-                arguments.SetInt(0, methodResult.ExecutionId);
+                arguments.SetInt64(0, methodResult.ExecutionId);
                 arguments.SetBool(1, methodResult.Success);
                 arguments.SetString(2, methodResult.Error);
                 if (methodResult.Result != null)
@@ -90,7 +90,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
             {
                 message = CefProcessMessage.Create(Messages.DynamicObjectResultMessageName);
                 var arguments = message.Arguments;
-                arguments.SetInt(0, dynamicObjectResponse.ExecutionId);
+                arguments.SetInt64(0, dynamicObjectResponse.ExecutionId);
                 arguments.SetBool(1, dynamicObjectResponse.Success);
                 arguments.SetString(2, dynamicObjectResponse.Exception);
 
@@ -107,8 +107,8 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
             {
                 message = CefProcessMessage.Create(Messages.CallbackExecutionMessageName);
                 var arguments = message.Arguments;
-                arguments.SetInt(0, callbackExecution.ExecutionId);
-                arguments.SetInt(1, callbackExecution.FunctionId);
+                arguments.SetInt64(0, callbackExecution.ExecutionId);
+                arguments.SetInt64(1, callbackExecution.FunctionId);
                 using (var argList = CefListValue.Create())
                 {
                     for (var i = 0; i < callbackExecution.Parameters.Length; i++)
@@ -123,7 +123,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
             {
                 message = CefProcessMessage.Create(Messages.DeleteCallbackMessageName);
                 var arguments = message.Arguments;
-                arguments.SetInt(0, deleteCallback.FunctionId);
+                arguments.SetInt64(0, deleteCallback.FunctionId);
             }
 
             return message;
@@ -139,7 +139,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                 case Messages.DynamicObjectResultMessageName:
                     var dynamicResult = new DynamicObjectResponse
                     {
-                        ExecutionId = args.GetInt(0),
+                        ExecutionId = args.GetInt64(0),
                         Success = args.GetBool(1),
                     };
 
@@ -164,7 +164,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                 case Messages.MethodResultMessageName:
                     var methodResult = new MethodResult<CefValue>()
                     {
-                        ExecutionId = args.GetInt(0),
+                        ExecutionId = args.GetInt64(0),
                         Success = args.GetBool(1),
                         Error = args.GetString(2),
                         Result = args.GetValue(3)
@@ -186,8 +186,8 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
 
                         var callbackExecution = new CallbackExecution<CefValue>
                         {
-                            ExecutionId = args.GetInt(0),
-                            FunctionId = args.GetInt(1),
+                            ExecutionId = args.GetInt64(0),
+                            FunctionId = args.GetInt64(1),
                             Parameters = paramList
                         };
 
@@ -198,7 +198,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                     }
                     break;
                 case Messages.DeleteCallbackMessageName:
-                    var functionId = args.GetInt(0);
+                    var functionId = args.GetInt64(0);
 
                     result = new RpcRequest<CefValue>
                     {
@@ -224,9 +224,9 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                 case Messages.MethodExecutionMessageName:
                     var execution = new MethodExecution<CefValue>
                     {
-                        ExecutionId = args.GetInt(0),
-                        MethodId = args.GetInt(1),
-                        ObjectId = args.GetInt(2)
+                        ExecutionId = args.GetInt64(0),
+                        MethodId = args.GetInt64(1),
+                        ObjectId = args.GetInt64(2)
                     };
                     var parameters = args.GetList(3);
                     var paramValues = new List<CefValue>();
@@ -244,7 +244,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                     break;
                 case Messages.DynamicObjectRequestMessageName:
                     var req = new DynamicObjectRequest();
-                    req.ExecutionId = args.GetInt(0);
+                    req.ExecutionId = args.GetInt64(0);
                     req.Name = args.GetString(1);
 
                     result = new RpcResponse<CefValue>
@@ -255,7 +255,7 @@ namespace DSerfozo.RpcBindings.CefGlue.Common.Serialization
                 case Messages.CallbackResultMessageName:
                     var resp = new CallbackResult<CefValue>();
 
-                    resp.ExecutionId = args.GetInt(0);
+                    resp.ExecutionId = args.GetInt64(0);
                     resp.Success = args.GetBool(1);
                     if (!resp.Success)
                     {
