+++ src/DSerfozo.RpcBindings.Json/Model/CallbackParameter.cs
@@ -5,6 +5,6 @@ namespace DSerfozo.RpcBindings.Json.Model
     public class CallbackParameter
     {
         [JsonProperty("functionId")]
-        public int FunctionId { get; set; }
+        public long FunctionId { get; set; }
     }
 }
