+++ src/DSerfozo.RpcBindings/Execution/PropertyExecutor.cs
@@ -8,10 +8,10 @@ namespace DSerfozo.RpcBindings.Execution
 {
     public class PropertyExecutor<TMarshal> : IPropertyExecutor<TMarshal>
     {
-        private readonly IReadOnlyDictionary<int, ObjectDescriptor> objects;
+        private readonly IReadOnlyDictionary<long, ObjectDescriptor> objects;
         private readonly IParameterBinder<TMarshal> parameterBinder;
 
-        public PropertyExecutor(IReadOnlyDictionary<int, ObjectDescriptor> objects, IParameterBinder<TMarshal> parameterBinder)
+        public PropertyExecutor(IReadOnlyDictionary<long, ObjectDescriptor> objects, IParameterBinder<TMarshal> parameterBinder)
         {
             this.objects = objects;
             this.parameterBinder = parameterBinder;
