+++ src/DSerfozo.RpcBindings/Marshaling/CallbackFactory.cs
@@ -14,7 +14,7 @@ namespace DSerfozo.RpcBindings.Marshaling
             this.callbackExecutor = callbackExecutor;
         }
 
-        public object CreateCallback(int id, Type delegateType, IParameterBinder<TMarshal> parameterBinder)
+        public object CreateCallback(long id, Type delegateType, IParameterBinder<TMarshal> parameterBinder)
         {
             if(delegateType == null)
             {
@@ -26,12 +26,12 @@ namespace DSerfozo.RpcBindings.Marshaling
             }
         }
 
-        private Delegate Create(int id, Type delegateType, IParameterBinder<TMarshal> parameterBinder)
+        private Delegate Create(long id, Type delegateType, IParameterBinder<TMarshal> parameterBinder)
         {
             return generator.Generate(delegateType, id, callbackExecutor, parameterBinder);
         }
 
-        private ICallback Create(int id, IParameterBinder<TMarshal> parameterBinder)
+        private ICallback Create(long id, IParameterBinder<TMarshal> parameterBinder)
         {
             return new Callback<TMarshal>(id, callbackExecutor, parameterBinder);
         }
